Disambiguate header parsing
The definition of the Cross-Origin-Embedder-Policy currently includes the
following text:

> In order to support forward-compatibility with as-yet-unknown request
> types, user agents MUST ignore [the Cross-Origin-Embedder-Policy HTTP
> Response Header] if it contains an invalid value. Likewise, user
> agents MUST ignore this header if the value cannot be parsed as a
> sh-token.

However, the subsequently-defined parsing algorithm recognize any "item"
value [1]. As a further complication, the Structured Header parsing
algorithm is misused: this proposal interprets the return value as a
string, but the return value is actually a tuple consisting of a string
and a map. That makes it unclear if/how parameters should influence the
result.

value                   | header definition  | parsing algorithm
------------------------|--------------------|------------------
`require-corp`          | `require-corp`     | `require-corp`
`"require-corp"`        | `none`             | `require-corp`
`*cmVxdWlyZS1jb3Jw*`    | `none`             | `require-corp`
`require-corp;`         | `none`             | ?
`require-corp; foo=bar` | `none`             | ?

Update the parsing algorithm to align with the description of the
header.

[1] https://tools.ietf.org/html/draft-ietf-httpbis-header-structure-14#section-3.3